requirements.txt
pyusb
tornado
